+++ packages/hash/tests/unit/go.spec.ts
@@ -1,5 +1,5 @@
 import "jest";
-import { Hash, POP, PUSH, REPLACE } from "../../src";
+import { Hash } from "../../src";
 import { JSDOM } from "jsdom";
 
 function ignoreFirstCall(fn) {
@@ -55,15 +55,15 @@ describe("go", () => {
         pathname: "/four",
         key: "3.0"
       });
-      expect(pending.action).toBe(POP);
+      expect(pending.action).toBe("pop");
       done();
     }
     testHistory.respondWith(router);
-    testHistory.navigate("/two", PUSH); // 1.0
-    testHistory.navigate("/three", PUSH); // 2.0
-    testHistory.navigate("/four", PUSH); // 3.0
-    testHistory.navigate("/five", PUSH); // 4.0
-    testHistory.navigate("/six", PUSH); // 5.0
+    testHistory.navigate("/two", "push"); // 1.0
+    testHistory.navigate("/three", "push"); // 2.0
+    testHistory.navigate("/four", "push"); // 3.0
+    testHistory.navigate("/five", "push"); // 4.0
+    testHistory.navigate("/six", "push"); // 5.0
     setup = true;
     testHistory.go(-2);
   });
@@ -83,10 +83,10 @@ describe("go", () => {
           cancelGo = pending.cancel;
           // trigger a push call and don't resolve the go
           testHistory.respondWith(pushRouter);
-          testHistory.navigate("/seven", PUSH);
+          testHistory.navigate("/seven", "push");
         });
         const pushRouter = ignoreFirstCall(function(pending) {
-          cancelGo(PUSH);
+          cancelGo("push");
           setTimeout(() => {
             expect(window.location.hash).toBe("#/six");
             done();
@@ -94,11 +94,11 @@ describe("go", () => {
         });
 
         testHistory.respondWith(initialRouter);
-        testHistory.navigate("/two", PUSH); // 1.0
-        testHistory.navigate("/three", PUSH); // 2.0
-        testHistory.navigate("/four", PUSH); // 3.0
-        testHistory.navigate("/five", PUSH); // 4.0
-        testHistory.navigate("/six", PUSH); // 5.0
+        testHistory.navigate("/two", "push"); // 1.0
+        testHistory.navigate("/three", "push"); // 2.0
+        testHistory.navigate("/four", "push"); // 3.0
+        testHistory.navigate("/five", "push"); // 4.0
+        testHistory.navigate("/six", "push"); // 5.0
 
         testHistory.respondWith(goRouter);
         testHistory.go(-2);
@@ -117,10 +117,10 @@ describe("go", () => {
           cancelGo = pending.cancel;
           // trigger a push call and don't resolve the go
           testHistory.respondWith(replaceRouter);
-          testHistory.navigate("/seven", REPLACE);
+          testHistory.navigate("/seven", "replace");
         });
         const replaceRouter = ignoreFirstCall(function(pending) {
-          cancelGo(REPLACE);
+          cancelGo("replace");
           setTimeout(() => {
             expect(window.location.hash).toBe("#/six");
             done();
@@ -128,11 +128,11 @@ describe("go", () => {
         });
 
         testHistory.respondWith(initialRouter);
-        testHistory.navigate("/two", PUSH); // 1.0
-        testHistory.navigate("/three", PUSH); // 2.0
-        testHistory.navigate("/four", PUSH); // 3.0
-        testHistory.navigate("/five", PUSH); // 4.0
-        testHistory.navigate("/six", PUSH); // 5.0
+        testHistory.navigate("/two", "push"); // 1.0
+        testHistory.navigate("/three", "push"); // 2.0
+        testHistory.navigate("/four", "push"); // 3.0
+        testHistory.navigate("/five", "push"); // 4.0
+        testHistory.navigate("/six", "push"); // 5.0
 
         testHistory.respondWith(goRouter);
         testHistory.go(-2);
@@ -154,7 +154,7 @@ describe("go", () => {
           testHistory.go(-1);
         });
         const popRouter = ignoreFirstCall(function(pending) {
-          cancelGo(POP);
+          cancelGo("pop");
           setTimeout(() => {
             expect(window.location.hash).toBe("#/three");
             done();
@@ -162,11 +162,11 @@ describe("go", () => {
         });
 
         testHistory.respondWith(initialRouter);
-        testHistory.navigate("/two", PUSH); // 1.0
-        testHistory.navigate("/three", PUSH); // 2.0
-        testHistory.navigate("/four", PUSH); // 3.0
-        testHistory.navigate("/five", PUSH); // 4.0
-        testHistory.navigate("/six", PUSH); // 5.0
+        testHistory.navigate("/two", "push"); // 1.0
+        testHistory.navigate("/three", "push"); // 2.0
+        testHistory.navigate("/four", "push"); // 3.0
+        testHistory.navigate("/five", "push"); // 4.0
+        testHistory.navigate("/six", "push"); // 5.0
 
         testHistory.respondWith(goRouter);
         testHistory.go(-2);
@@ -195,8 +195,8 @@ describe("go", () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router);
 
-      testHistory.navigate("/two", PUSH); // 1.0
-      testHistory.navigate("/three", PUSH); // 2.0
+      testHistory.navigate("/two", "push"); // 1.0
+      testHistory.navigate("/three", "push"); // 2.0
       setup = true;
       testHistory.go(-2);
     });
@@ -211,8 +211,8 @@ describe("go", () => {
       }
       testHistory.respondWith(router);
 
-      testHistory.navigate("/two", PUSH); // 1.0
-      testHistory.navigate("/three", PUSH); // 2.0
+      testHistory.navigate("/two", "push"); // 1.0
+      testHistory.navigate("/three", "push"); // 2.0
       // don't add function until we have setup the history
       testHistory.confirmWith(confirm);
       testHistory.go(-2);