+++ packages/root/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./types";
 
 import Common from "./common";
-import { PUSH, REPLACE, ANCHOR, POP } from "./constants";
-export { Common, PUSH, REPLACE, ANCHOR, POP };
+export { Common };